Rename dukeapp class to duke
