+++ README.md
@@ -17,7 +17,7 @@ Prerequisites: JDK 11, update Intellij to the most recent version.
    in [here](https://www.jetbrains.com/help/idea/sdk.html#set-up-jdk).<br>
    In the same dialog, set the **Project language level** field to
    the `SDK default` option.
-3. After that, locate the `src/main/java/dukeapp/Duke.java` file, right-click
+3. After that, locate the `src/main/java/duke/Duke.java` file, right-click
    it, and choose `Run Duke.main()` (if the code editor is showing
    compile errors, try restarting the IDE). If the setup is correct, you should
    see something like the below as the output:
