+++ build.gradle
@@ -28,7 +28,7 @@ test {
 }
 
 application {
-    mainClass.set("dukeapp.Duke")
+    mainClass.set("duke.Duke")
 }
 
 shadowJar {
