+++ src/main/java/duke/commands/DeleteCommand.java
@@ -1,19 +1,19 @@
-package dukeapp.commands;
+package duke.commands;
 
-import dukeapp.Messages;
-import dukeapp.TaskList;
-import dukeapp.Ui;
+import duke.Messages;
+import duke.TaskList;
+import duke.Ui;
 
 import java.io.IOException;
 
 /**
- * Unmarks a task.
+ * Deletes a task.
  */
-public class UnmarkCommand implements Command {
+public class DeleteCommand implements Command {
     /**
-     * Sets a given task to be unmarked.
+     * Deletes a task from the application.
      *
-     * @param input    The user input of the task to be unmarked.
+     * @param input    The user input of the task to be deleted.
      * @param taskList The application's task list.
      * @param ui       The UI of the application.
      */
@@ -23,10 +23,12 @@ public class UnmarkCommand implements Command {
         String indexString = args[1];
         int index = Integer.parseInt(indexString) - 1;
         try {
-            String taskString = taskList.unmarkTask(index);
-            ui.displayMessage(String.format(Messages.UNMARKED_MESSAGE, taskString));
+            String taskString = taskList.deleteTask(index);
+            ui.displayMessage(String.format(Messages.DELETE_MESSAGE, taskString,
+                    taskList.getTaskCount()));
         } catch (IOException e) {
             ui.displayError(e.getMessage());
         }
+
     }
 }
