+++ src/main/java/duke/commands/MarkCommand.java
@@ -1,19 +1,19 @@
-package dukeapp.commands;
+package duke.commands;
 
-import dukeapp.Messages;
-import dukeapp.TaskList;
-import dukeapp.Ui;
+import duke.Messages;
+import duke.TaskList;
+import duke.Ui;
 
 import java.io.IOException;
 
 /**
- * Deletes a task.
+ * Marks a task.
  */
-public class DeleteCommand implements Command {
+public class MarkCommand implements Command {
     /**
-     * Deletes a task from the application.
+     * Sets a given task to be marked.
      *
-     * @param input    The user input of the task to be deleted.
+     * @param input    The user input of the task to be marked.
      * @param taskList The application's task list.
      * @param ui       The UI of the application.
      */
@@ -23,12 +23,10 @@ public class DeleteCommand implements Command {
         String indexString = args[1];
         int index = Integer.parseInt(indexString) - 1;
         try {
-            String taskString = taskList.deleteTask(index);
-            ui.displayMessage(String.format(Messages.DELETE_MESSAGE, taskString,
-                    taskList.getTaskCount()));
+            String taskString = taskList.markTask(index);
+            ui.displayMessage(String.format(Messages.MARKED_MESSAGE, taskString));
         } catch (IOException e) {
             ui.displayError(e.getMessage());
         }
-
     }
 }
