--- text-ui-test/data/duke.txt
@@ -1,12 +1,0 @@
-D | 0 | return book | 2023-09-30 1200
-E | 0 | project meeting | 2023-08-03 1430-2023-08-30 1630
-D | 0 | ma2001 hw | 2023-08-03 0000
-D | 0 | ma1521 hw | 2023-08-03 0100
-E | 0 | code | 2023-08-03 0000-2023-08-03 0030
-E | 0 | run | 2023-08-03 0000-2023-08-03 0045
-E | 0 | yoga | 2023-08-03 0000-2023-08-03 0145
-E | 0 | nap | 2023-08-03 0000-2023-08-03 0230
-E | 0 | dorm workshop | 2023-08-03 0000-2023-08-03 0315
-E | 0 | walk around nus | 2023-08-03 0000-2023-08-03 0400
-E | 0 | buy supper | 2023-08-03 0000-2023-08-03 0445
-E | 0 | sleep | 2023-08-03 0000-2023-08-03 0530
