+++ text-ui-test/runtest.sh
@@ -20,14 +20,14 @@ then
 fi
 
 # compile the code into the bin folder, terminates if error occurred
-if ! javac -encoding UTF-8 -cp ../src/main/java -Xlint:none -d ../bin ../src/main/java/dukeapp/*.java
+if ! javac -encoding UTF-8 -cp ../src/main/java -Xlint:none -d ../bin ../src/main/java/duke/*.java
 then
     echo "********** BUILD FAILURE **********"
     exit 1
 fi
 
 # run the program, feed commands from input.txt file and redirect the output to the ACTUAL.TXT
-java -classpath ../bin dukeapp.Duke < input.txt > ACTUAL.TXT
+java -classpath ../bin duke.Duke < input.txt > ACTUAL.TXT
 
 # convert to UNIX format
 cp EXPECTED.TXT EXPECTED-UNIX.TXT
